+++ src/OpenTelemetry/Metrics/MetricPoint.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using OpenTelemetry.Internal;
 
@@ -19,11 +18,6 @@ public struct MetricPoint
     // but we always decrement it (in the Update methods). This should be fine.
     // ReferenceCount doesn't matter for MetricPoint with no tags and overflow attribute as they are never reclaimed.
     internal int ReferenceCount;
-
-    // When the AggregatorStore is reclaiming MetricPoints, this serves the purpose of validating the a given thread is using the right
-    // MetricPoint for update by checking it against what as added in the Dictionary. Also, when a thread finds out that the MetricPoint
-    // that its using is already reclaimed, this helps avoid sorting of the tags for adding a new Dictionary entry.
-    internal LookupData? LookupData;
 
     private const int DefaultSimpleReservoirPoolSize = 1;
 
@@ -155,6 +149,12 @@ public struct MetricPoint
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private set;
     }
+
+    // When the AggregatorStore is reclaiming MetricPoints, this serves the purpose of validating the a given thread is using the right
+    // MetricPoint for update by checking it against what as added in the Dictionary. Also, when a thread finds out that the MetricPoint
+    // that its using is already reclaimed, this helps avoid sorting of the tags for adding a new Dictionary entry.
+    // Snapshot method can use this to skip trying to reclaim indices which have already been reclaimed and added to the queue.
+    internal LookupData? LookupData { readonly get; private set; }
 
     internal readonly bool IsInitialized => this.aggregatorStore != null;
 
@@ -373,10 +373,10 @@ public struct MetricPoint
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     internal
 #endif
-        readonly bool TryGetExemplars([NotNullWhen(true)] out ReadOnlyExemplarCollection? exemplars)
+        readonly bool TryGetExemplars(out ReadOnlyExemplarCollection exemplars)
     {
-        exemplars = this.mpComponents?.Exemplars;
-        return exemplars.HasValue;
+        exemplars = this.mpComponents?.Exemplars ?? ReadOnlyExemplarCollection.Empty;
+        return exemplars.MaximumCount > 0;
     }
 
     internal readonly MetricPoint Copy()
@@ -463,15 +463,7 @@ public struct MetricPoint
                         this.runningValue.AsLong += number;
                     }
 
-                    if (isSampled)
-                    {
-                        Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-                        // TODO: Need to ensure that the lock is always released.
-                        // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-                        this.mpComponents.ExemplarReservoir!.Offer(
-                            new ExemplarMeasurement<long>(number, tags));
-                    }
+                    this.OfferExemplarIfSampled(number, tags, isSampled);
 
                     this.mpComponents.ReleaseLock();
 
@@ -484,15 +476,7 @@ public struct MetricPoint
 
                     this.runningValue.AsLong = number;
 
-                    if (isSampled)
-                    {
-                        Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-                        // TODO: Need to ensure that the lock is always released.
-                        // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-                        this.mpComponents.ExemplarReservoir!.Offer(
-                            new ExemplarMeasurement<long>(number, tags));
-                    }
+                    this.OfferExemplarIfSampled(number, tags, isSampled);
 
                     this.mpComponents.ReleaseLock();
 
@@ -505,15 +489,7 @@ public struct MetricPoint
 
                     this.runningValue.AsLong = number;
 
-                    if (isSampled)
-                    {
-                        Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-                        // TODO: Need to ensure that the lock is always released.
-                        // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-                        this.mpComponents.ExemplarReservoir!.Offer(
-                            new ExemplarMeasurement<long>(number, tags));
-                    }
+                    this.OfferExemplarIfSampled(number, tags, isSampled);
 
                     this.mpComponents.ReleaseLock();
 
@@ -522,37 +498,37 @@ public struct MetricPoint
 
             case AggregationType.Histogram:
                 {
-                    this.UpdateHistogram((double)number, tags, reportExemplar: true, isSampled);
+                    this.UpdateHistogram((double)number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.HistogramWithMinMax:
                 {
-                    this.UpdateHistogramWithMinMax((double)number, tags, reportExemplar: true, isSampled);
+                    this.UpdateHistogramWithMinMax((double)number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.HistogramWithBuckets:
                 {
-                    this.UpdateHistogramWithBuckets((double)number, tags, reportExemplar: true, isSampled);
+                    this.UpdateHistogramWithBuckets((double)number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.HistogramWithMinMaxBuckets:
                 {
-                    this.UpdateHistogramWithBucketsAndMinMax((double)number, tags, reportExemplar: true, isSampled);
+                    this.UpdateHistogramWithBucketsAndMinMax((double)number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.Base2ExponentialHistogram:
                 {
-                    this.UpdateBase2ExponentialHistogram((double)number, tags, reportExemplar: true, isSampled);
+                    this.UpdateBase2ExponentialHistogram((double)number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.Base2ExponentialHistogramWithMinMax:
                 {
-                    this.UpdateBase2ExponentialHistogramWithMinMax((double)number, tags, reportExemplar: true, isSampled);
+                    this.UpdateBase2ExponentialHistogramWithMinMax((double)number, tags, isSampled);
                     break;
                 }
         }
@@ -637,15 +613,7 @@ public struct MetricPoint
                         this.runningValue.AsDouble += number;
                     }
 
-                    if (isSampled)
-                    {
-                        Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-                        // TODO: Need to ensure that the lock is always released.
-                        // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-                        this.mpComponents.ExemplarReservoir!.Offer(
-                            new ExemplarMeasurement<double>(number, tags));
-                    }
+                    this.OfferExemplarIfSampled(number, tags, isSampled);
 
                     this.mpComponents.ReleaseLock();
 
@@ -661,15 +629,7 @@ public struct MetricPoint
                         this.runningValue.AsDouble = number;
                     }
 
-                    if (isSampled)
-                    {
-                        Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-                        // TODO: Need to ensure that the lock is always released.
-                        // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-                        this.mpComponents.ExemplarReservoir!.Offer(
-                            new ExemplarMeasurement<double>(number, tags));
-                    }
+                    this.OfferExemplarIfSampled(number, tags, isSampled);
 
                     this.mpComponents.ReleaseLock();
 
@@ -685,15 +645,7 @@ public struct MetricPoint
                         this.runningValue.AsDouble = number;
                     }
 
-                    if (isSampled)
-                    {
-                        Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-                        // TODO: Need to ensure that the lock is always released.
-                        // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-                        this.mpComponents.ExemplarReservoir!.Offer(
-                            new ExemplarMeasurement<double>(number, tags));
-                    }
+                    this.OfferExemplarIfSampled(number, tags, isSampled);
 
                     this.mpComponents.ReleaseLock();
 
@@ -702,37 +654,37 @@ public struct MetricPoint
 
             case AggregationType.Histogram:
                 {
-                    this.UpdateHistogram(number, tags, reportExemplar: true, isSampled);
+                    this.UpdateHistogram(number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.HistogramWithMinMax:
                 {
-                    this.UpdateHistogramWithMinMax(number, tags, reportExemplar: true, isSampled);
+                    this.UpdateHistogramWithMinMax(number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.HistogramWithBuckets:
                 {
-                    this.UpdateHistogramWithBuckets(number, tags, reportExemplar: true, isSampled);
+                    this.UpdateHistogramWithBuckets(number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.HistogramWithMinMaxBuckets:
                 {
-                    this.UpdateHistogramWithBucketsAndMinMax(number, tags, reportExemplar: true, isSampled);
+                    this.UpdateHistogramWithBucketsAndMinMax(number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.Base2ExponentialHistogram:
                 {
-                    this.UpdateBase2ExponentialHistogram(number, tags, reportExemplar: true, isSampled);
+                    this.UpdateBase2ExponentialHistogram(number, tags, isSampled);
                     break;
                 }
 
             case AggregationType.Base2ExponentialHistogramWithMinMax:
                 {
-                    this.UpdateBase2ExponentialHistogramWithMinMax(number, tags, reportExemplar: true, isSampled);
+                    this.UpdateBase2ExponentialHistogramWithMinMax(number, tags, isSampled);
                     break;
                 }
         }
@@ -993,13 +945,14 @@ public struct MetricPoint
     internal void TakeSnapshotWithExemplar(bool outputDelta)
     {
         Debug.Assert(this.mpComponents != null, "this.mpComponents was null");
+        Debug.Assert(this.mpComponents!.ExemplarReservoir != null, "this.mpComponents.ExemplarReservoir was null");
 
         switch (this.aggType)
         {
             case AggregationType.LongSumIncomingDelta:
             case AggregationType.LongSumIncomingCumulative:
                 {
-                    this.mpComponents!.AcquireLock();
+                    this.mpComponents.AcquireLock();
 
                     if (outputDelta)
                     {
@@ -1013,7 +966,7 @@ public struct MetricPoint
                         this.snapshotValue.AsLong = this.runningValue.AsLong;
                     }
 
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
 
                     this.mpComponents.ReleaseLock();
 
@@ -1037,7 +990,7 @@ public struct MetricPoint
                         this.snapshotValue.AsDouble = this.runningValue.AsDouble;
                     }
 
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
 
                     this.mpComponents.ReleaseLock();
 
@@ -1046,11 +999,11 @@ public struct MetricPoint
 
             case AggregationType.LongGauge:
                 {
-                    this.mpComponents!.AcquireLock();
+                    this.mpComponents.AcquireLock();
 
                     this.snapshotValue.AsLong = this.runningValue.AsLong;
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
 
                     this.mpComponents.ReleaseLock();
 
@@ -1059,11 +1012,11 @@ public struct MetricPoint
 
             case AggregationType.DoubleGauge:
                 {
-                    this.mpComponents!.AcquireLock();
+                    this.mpComponents.AcquireLock();
 
                     this.snapshotValue.AsDouble = this.runningValue.AsDouble;
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
 
                     this.mpComponents.ReleaseLock();
 
@@ -1072,9 +1025,9 @@ public struct MetricPoint
 
             case AggregationType.HistogramWithBuckets:
                 {
-                    Debug.Assert(this.mpComponents!.HistogramBuckets != null, "HistogramBuckets was null");
+                    Debug.Assert(this.mpComponents.HistogramBuckets != null, "HistogramBuckets was null");
 
-                    var histogramBuckets = this.mpComponents!.HistogramBuckets!;
+                    var histogramBuckets = this.mpComponents.HistogramBuckets!;
 
                     this.mpComponents.AcquireLock();
 
@@ -1089,7 +1042,7 @@ public struct MetricPoint
 
                     histogramBuckets.Snapshot(outputDelta);
 
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
 
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
@@ -1100,9 +1053,9 @@ public struct MetricPoint
 
             case AggregationType.Histogram:
                 {
-                    Debug.Assert(this.mpComponents!.HistogramBuckets != null, "HistogramBuckets was null");
+                    Debug.Assert(this.mpComponents.HistogramBuckets != null, "HistogramBuckets was null");
 
-                    var histogramBuckets = this.mpComponents!.HistogramBuckets!;
+                    var histogramBuckets = this.mpComponents.HistogramBuckets!;
 
                     this.mpComponents.AcquireLock();
 
@@ -1115,7 +1068,7 @@ public struct MetricPoint
                         histogramBuckets.RunningSum = 0;
                     }
 
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                     this.mpComponents.ReleaseLock();
@@ -1125,9 +1078,9 @@ public struct MetricPoint
 
             case AggregationType.HistogramWithMinMaxBuckets:
                 {
-                    Debug.Assert(this.mpComponents!.HistogramBuckets != null, "HistogramBuckets was null");
+                    Debug.Assert(this.mpComponents.HistogramBuckets != null, "HistogramBuckets was null");
 
-                    var histogramBuckets = this.mpComponents!.HistogramBuckets!;
+                    var histogramBuckets = this.mpComponents.HistogramBuckets!;
 
                     this.mpComponents.AcquireLock();
 
@@ -1146,7 +1099,7 @@ public struct MetricPoint
 
                     histogramBuckets.Snapshot(outputDelta);
 
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                     this.mpComponents.ReleaseLock();
@@ -1156,9 +1109,9 @@ public struct MetricPoint
 
             case AggregationType.HistogramWithMinMax:
                 {
-                    Debug.Assert(this.mpComponents!.HistogramBuckets != null, "HistogramBuckets was null");
+                    Debug.Assert(this.mpComponents.HistogramBuckets != null, "HistogramBuckets was null");
 
-                    var histogramBuckets = this.mpComponents!.HistogramBuckets!;
+                    var histogramBuckets = this.mpComponents.HistogramBuckets!;
 
                     this.mpComponents.AcquireLock();
 
@@ -1175,7 +1128,7 @@ public struct MetricPoint
                         histogramBuckets.RunningMax = double.NegativeInfinity;
                     }
 
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                     this.mpComponents.ReleaseLock();
@@ -1185,9 +1138,9 @@ public struct MetricPoint
 
             case AggregationType.Base2ExponentialHistogram:
                 {
-                    Debug.Assert(this.mpComponents!.Base2ExponentialBucketHistogram != null, "Base2ExponentialBucketHistogram was null");
+                    Debug.Assert(this.mpComponents.Base2ExponentialBucketHistogram != null, "Base2ExponentialBucketHistogram was null");
 
-                    var histogram = this.mpComponents!.Base2ExponentialBucketHistogram!;
+                    var histogram = this.mpComponents.Base2ExponentialBucketHistogram!;
 
                     this.mpComponents.AcquireLock();
 
@@ -1202,7 +1155,7 @@ public struct MetricPoint
                         histogram.Reset();
                     }
 
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                     this.mpComponents.ReleaseLock();
@@ -1212,9 +1165,9 @@ public struct MetricPoint
 
             case AggregationType.Base2ExponentialHistogramWithMinMax:
                 {
-                    Debug.Assert(this.mpComponents!.Base2ExponentialBucketHistogram != null, "Base2ExponentialBucketHistogram was null");
+                    Debug.Assert(this.mpComponents.Base2ExponentialBucketHistogram != null, "Base2ExponentialBucketHistogram was null");
 
-                    var histogram = this.mpComponents!.Base2ExponentialBucketHistogram!;
+                    var histogram = this.mpComponents.Base2ExponentialBucketHistogram!;
 
                     this.mpComponents.AcquireLock();
 
@@ -1233,7 +1186,7 @@ public struct MetricPoint
                         histogram.RunningMax = double.NegativeInfinity;
                     }
 
-                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir?.Collect();
+                    this.mpComponents.Exemplars = this.mpComponents.ExemplarReservoir!.Collect();
                     this.MetricPointStatus = MetricPointStatus.NoCollectPending;
 
                     this.mpComponents.ReleaseLock();
@@ -1243,7 +1196,16 @@ public struct MetricPoint
         }
     }
 
-    private void UpdateHistogram(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false, bool isSampled = false)
+    /// <summary>
+    /// Denote that this MetricPoint is reclaimed.
+    /// </summary>
+    internal void Reclaim()
+    {
+        this.LookupData = null;
+        this.mpComponents = null;
+    }
+
+    private void UpdateHistogram(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool isSampled = false)
     {
         Debug.Assert(this.mpComponents?.HistogramBuckets != null, "HistogramBuckets was null");
 
@@ -1257,20 +1219,12 @@ public struct MetricPoint
             histogramBuckets.RunningSum += number;
         }
 
-        if (reportExemplar && isSampled)
-        {
-            Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-            // TODO: Need to ensure that the lock is always released.
-            // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-            this.mpComponents.ExemplarReservoir!.Offer(
-                new ExemplarMeasurement<double>(number, tags));
-        }
+        this.OfferExemplarIfSampled(number, tags, isSampled);
 
         this.mpComponents.ReleaseLock();
     }
 
-    private void UpdateHistogramWithMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false, bool isSampled = false)
+    private void UpdateHistogramWithMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool isSampled = false)
     {
         Debug.Assert(this.mpComponents?.HistogramBuckets != null, "HistogramBuckets was null");
 
@@ -1286,26 +1240,18 @@ public struct MetricPoint
             histogramBuckets.RunningMax = Math.Max(histogramBuckets.RunningMax, number);
         }
 
-        if (reportExemplar && isSampled)
-        {
-            Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-            // TODO: Need to ensure that the lock is always released.
-            // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-            this.mpComponents.ExemplarReservoir!.Offer(
-                new ExemplarMeasurement<double>(number, tags));
-        }
+        this.OfferExemplarIfSampled(number, tags, isSampled);
 
         this.mpComponents.ReleaseLock();
     }
 
-    private void UpdateHistogramWithBuckets(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false, bool isSampled = false)
+    private void UpdateHistogramWithBuckets(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool isSampled = false)
     {
         Debug.Assert(this.mpComponents?.HistogramBuckets != null, "HistogramBuckets was null");
 
         var histogramBuckets = this.mpComponents!.HistogramBuckets;
 
-        int i = histogramBuckets!.FindBucketIndex(number);
+        int bucketIndex = histogramBuckets!.FindBucketIndex(number);
 
         this.mpComponents.AcquireLock();
 
@@ -1313,29 +1259,21 @@ public struct MetricPoint
         {
             this.runningValue.AsLong++;
             histogramBuckets.RunningSum += number;
-            histogramBuckets.BucketCounts[i].RunningValue++;
+            histogramBuckets.BucketCounts[bucketIndex].RunningValue++;
         }
 
-        if (reportExemplar && isSampled)
-        {
-            Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-            // TODO: Need to ensure that the lock is always released.
-            // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-            this.mpComponents.ExemplarReservoir!.Offer(
-                new ExemplarMeasurement<double>(number, tags, i));
-        }
+        this.OfferExplicitBucketHistogramExemplarIfSampled(number, tags, bucketIndex, isSampled);
 
         this.mpComponents.ReleaseLock();
     }
 
-    private void UpdateHistogramWithBucketsAndMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false, bool isSampled = false)
+    private void UpdateHistogramWithBucketsAndMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool isSampled = false)
     {
         Debug.Assert(this.mpComponents?.HistogramBuckets != null, "histogramBuckets was null");
 
         var histogramBuckets = this.mpComponents!.HistogramBuckets;
 
-        int i = histogramBuckets!.FindBucketIndex(number);
+        int bucketIndex = histogramBuckets!.FindBucketIndex(number);
 
         this.mpComponents.AcquireLock();
 
@@ -1343,26 +1281,18 @@ public struct MetricPoint
         {
             this.runningValue.AsLong++;
             histogramBuckets.RunningSum += number;
-            histogramBuckets.BucketCounts[i].RunningValue++;
+            histogramBuckets.BucketCounts[bucketIndex].RunningValue++;
 
             histogramBuckets.RunningMin = Math.Min(histogramBuckets.RunningMin, number);
             histogramBuckets.RunningMax = Math.Max(histogramBuckets.RunningMax, number);
         }
 
-        if (reportExemplar && isSampled)
-        {
-            Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-            // TODO: Need to ensure that the lock is always released.
-            // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-            this.mpComponents.ExemplarReservoir!.Offer(
-                new ExemplarMeasurement<double>(number, tags, i));
-        }
+        this.OfferExplicitBucketHistogramExemplarIfSampled(number, tags, bucketIndex, isSampled);
 
         this.mpComponents.ReleaseLock();
     }
 
-    private void UpdateBase2ExponentialHistogram(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false, bool isSampled = false)
+    private void UpdateBase2ExponentialHistogram(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool isSampled = false)
     {
         if (number < 0)
         {
@@ -1382,20 +1312,12 @@ public struct MetricPoint
             histogram.Record(number);
         }
 
-        if (reportExemplar && isSampled)
-        {
-            Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
-
-            // TODO: Need to ensure that the lock is always released.
-            // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-            this.mpComponents.ExemplarReservoir!.Offer(
-                new ExemplarMeasurement<double>(number, tags));
-        }
+        this.OfferExemplarIfSampled(number, tags, isSampled);
 
         this.mpComponents.ReleaseLock();
     }
 
-    private void UpdateBase2ExponentialHistogramWithMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool reportExemplar = false, bool isSampled = false)
+    private void UpdateBase2ExponentialHistogramWithMinMax(double number, ReadOnlySpan<KeyValuePair<string, object?>> tags = default, bool isSampled = false)
     {
         if (number < 0)
         {
@@ -1418,17 +1340,56 @@ public struct MetricPoint
             histogram.RunningMax = Math.Max(histogram.RunningMax, number);
         }
 
-        if (reportExemplar && isSampled)
+        this.OfferExemplarIfSampled(number, tags, isSampled);
+
+        this.mpComponents.ReleaseLock();
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private readonly void OfferExemplarIfSampled<T>(T number, ReadOnlySpan<KeyValuePair<string, object?>> tags, bool isSampled)
+        where T : struct
+    {
+        if (isSampled)
         {
-            Debug.Assert(this.mpComponents.ExemplarReservoir != null, "ExemplarReservoir was null");
+            Debug.Assert(this.mpComponents?.ExemplarReservoir != null, "ExemplarReservoir was null");
 
             // TODO: Need to ensure that the lock is always released.
             // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
-            this.mpComponents.ExemplarReservoir!.Offer(
-                new ExemplarMeasurement<double>(number, tags));
+            if (typeof(T) == typeof(long))
+            {
+                this.mpComponents!.ExemplarReservoir!.Offer(
+                    new ExemplarMeasurement<long>((long)(object)number, tags));
+            }
+            else if (typeof(T) == typeof(double))
+            {
+                this.mpComponents!.ExemplarReservoir!.Offer(
+                    new ExemplarMeasurement<double>((double)(object)number, tags));
+            }
+            else
+            {
+                Debug.Fail("Unexpected type");
+                this.mpComponents!.ExemplarReservoir!.Offer(
+                    new ExemplarMeasurement<double>(Convert.ToDouble(number), tags));
+            }
         }
+    }
 
-        this.mpComponents.ReleaseLock();
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private readonly void OfferExplicitBucketHistogramExemplarIfSampled(
+        double number,
+        ReadOnlySpan<KeyValuePair<string, object?>> tags,
+        int bucketIndex,
+        bool isSampled)
+    {
+        if (isSampled)
+        {
+            Debug.Assert(this.mpComponents?.ExemplarReservoir != null, "ExemplarReservoir was null");
+
+            // TODO: Need to ensure that the lock is always released.
+            // A custom implementation of `ExemplarReservoir.Offer` might throw an exception.
+            this.mpComponents!.ExemplarReservoir!.Offer(
+                new ExemplarMeasurement<double>(number, tags, bucketIndex));
+        }
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]