+++ src/OpenTelemetry/Metrics/ThreadStaticStorage.cs
@@ -1,6 +1,9 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+#if NET8_0_OR_GREATER
+using System.Collections.Frozen;
+#endif
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using OpenTelemetry.Internal;
@@ -54,7 +57,11 @@ internal sealed class ThreadStaticStorage
     internal void SplitToKeysAndValues(
         ReadOnlySpan<KeyValuePair<string, object?>> tags,
         int tagLength,
+#if NET8_0_OR_GREATER
+        FrozenSet<string> tagKeysInteresting,
+#else
         HashSet<string> tagKeysInteresting,
+#endif
         out KeyValuePair<string, object?>[]? tagKeysAndValues,
         out int actualLength)
     {